Update the Java code samples
This PR updates the code sample according to the decision made
with the documentation and the APIs team, to provide consistent
code samples across all clients.

Instead of populating a map with random entries all the time,
the default code sample to run now populates a map with HJV
entries so that it can be queryable. The part that queries the
map will be handled by the documentation team in their tutorials.

This PR also contains some small changes to the code sample like
getting rid of unnecessary abstractions, fixing the formatting, making
create mapping statements more friendly to non-java developers, etc.